Restructure folder structure according to python packaging
