+++ setup.py
@@ -1,5 +1,5 @@
 import os
-from setuptools import setup, find_packages
+from setuptools import setup, find_packages, Extension
 from version import version
 
 project_path = os.path.dirname(os.path.realpath(__file__))
@@ -12,9 +12,8 @@ install_requires = [x.strip() for x in content]
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
-
 setup(
-    name="src",
+    name="python_ms_core",
     version=version,
     author="Sujata Misra",
     author_email="sujatam@gaussiansolutions.com",
@@ -23,10 +22,19 @@ setup(
     long_description_content_type="text/markdown",
     url="https://github.com/TaskarCenterAtUW/TDEI-Python-ms-core",
     install_requires=install_requires,
-    packages=find_packages(),
+    packages=find_packages(where='src'),
+    # packages=['python_ms_core'],
+    # namespace_packages=['python_ms_core'],
+    # packages={
+    #     "python_ms_core": "src"
+    # },
     classifiers=[
         "Programming Language :: Python :: 3",
+        "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ],
     python_requires='>=3.9',
+    # ext_modules=[Extension('python_ms_core.src', ['python_ms_core.src.main'])]
+    # py_modules=['python_ms_core'],
+    package_dir={'': 'src'},
 )
