+++ src/python_ms_core/core/logger/providers/azure_logger_config.py
