+++ src/python_ms_core/core/queue/models/queue_message.py
