+++ src/python_ms_core/core/resource_errors/errors.py
