+++ src/python_ms_core/core/storage/abstract/file_entity.py
