+++ src/python_ms_core/models/abstracts/iconfig.py
@@ -1,4 +1,4 @@
-from src.provider_types.provider import Provider
+from src.python_ms_core.provider_types.provider import Provider
 
 
 class IConfig:
